requirements.txt
# Railway-optimized requirements (Linux compatible) - Updated for Hume compatibility
# Core Django packages
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.5.0
django-debug-toolbar==4.2.0
drf-yasg==1.21.7
daphne==4.1.2
channels==4.1.0
channels-redis==4.2.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
django-allauth==0.63.6
djangorestframework-simplejwt==5.3.0
cryptography==42.0.8

# API & WebSocket
hume==0.13.4
websockets==13.1
aiohttp==3.9.5
requests==2.31.0
beautifulsoup4==4.12.3

# Voice & Telephony
vonage==4.7.2
twilio==8.10.3

# Background Tasks & Caching
celery==5.3.4
django-celery-beat==2.6.0
django-celery-results==2.5.1
redis==5.0.7
django-redis==5.4.0

# File Handling
Pillow==10.4.0

# Utilities
python-dotenv==1.0.1
python-decouple==3.8
colorama==0.4.6
pytz==2024.1

# HTTP & Async
httpx==0.27.0
anyio==4.4.0
aiofiles==24.1.0

# Data Processing
pandas==2.1.4
numpy==1.24.4

# Email
django-ses==3.5.2

# Storage
boto3==1.34.131
django-storages==1.14.3

# Payment Processing
stripe==10.2.0

# Performance
whitenoise==6.6.0

# Monitoring
sentry-sdk==2.8.0

# Vector database and document processing
chromadb==0.4.24
PyPDF2==3.0.1
python-docx==1.1.0

# Audio processing
pydub==0.25.1

# Additional utilities
PyYAML==6.0.1
jsonschema==4.19.2